Feature/change jwt to jwt extended

+++ resources/item.py
@@ -1,73 +1,85 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import get_jwt_identity, jwt_required, get_jwt_claims, \
+    fresh_jwt_required, jwt_optional
 
 from models.item import ItemModel
 
-class Item(Resource):
-    TABLE_NAME = 'items'
 
+class Item(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument('price',
-        type=float,
-        required=True,
-        help="This field cannot be left blank!"
-    )
+                        type=float,
+                        required=True,
+                        help="This field cannot be left blank!"
+                        )
     parser.add_argument('store_id',
-        type=int,
-        required=True,
-        help="Every item needs a store id."
-    )
+                        type=int,
+                        required=True,
+                        help="Every item needs a store_id."
+                        )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
-            return item.in_json()
+            return item.json()
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
-            return {'message': "An item with name '{}' already exists.".format(name)}, 400
+            return {
+                'message': f"An item with name '{name}' already exists."
+            }, 400
 
-        data = Item.parser.parse_args()
+        data = self.parser.parse_args()
 
-        item = ItemModel(name, data['price'], data['store_id'])
+        item = ItemModel(name, **data)
 
         try:
             item.save_to_db()
         except:
-            return {"message": "An error occurred inserting the item."}, 500
+            return {
+                "message": "An error occurred while inserting the item."
+            }, 500
 
-        return item.in_json(), 201
+        return item.json(), 201
 
-
-
-    @jwt_required()
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
+            return {'message': 'Item deleted.'}
+        return {'message': 'Item not found.'}, 404
 
-        return {'message': 'Item deleted'}
-
-    @jwt_required()
     def put(self, name):
-        data = Item.parser.parse_args()
+        data = self.parser.parse_args()
+
         item = ItemModel.find_by_name(name)
 
-        if item is None:
-            item = ItemModel(name, data['price'], data['store_id'])
-        else:
+        if item:
             item.price = data['price']
-            item.store_id = data['store_id']
+        else:
+            item = ItemModel(name, **data)
 
         item.save_to_db()
-        return item.in_json()
+
+        return item.json()
 
 
 class ItemList(Resource):
-    TABLE_NAME = 'items'
-
+    @jwt_optional
     def get(self):
-        # return {'items': list(map(lambda x: x.in_json(), ItemModel.query.all()))}
-        return {'items': [item.in_json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+        }, 200
